Rework nova-common to use interface_ip
Currently the way that the various console options are being setup
does not work. It assumes an older structure for Ohai data that
doesn't really fit our Chef 11 use case. So instead of digging into
node attributes to find the IP address of a local interface, make use
of the openstack-common interface_ip function.