Remove inclusion of metadata and network recipes
There is no reason why the compute recipe should include the recipe
for the metadata service nor the nova-network service. This is
especially true for those not wanting the metadata service and/or
those using neutron.

In the case of network, this is included conditionally based upon
network.service_type, but there is really no reason this recipe should
not stand on its own.

Change-Id: I7545958f0ebfa667c96254f09b9c4c6f36c47b35